fix: address code review feedback
- JobsPromise now throws if error handler not set before adding promises
- Added clarifying comment explaining orphaned promise pattern in createAssembly
- WatchJobEmitter now properly cleans up file watchers on SIGINT/SIGTERM and errors
- Expanded --verbose/--quiet documentation with output level table

+++ src/cli/JobsPromise.ts
@@ -19,12 +19,17 @@ export default class JobsPromise {
   /**
    * Add a promise to track. If the promise rejects,
    * the error handler will be called.
+   * @throws Error if error handler has not been set via setErrorHandler()
    */
   add(promise: Promise<unknown>): void {
+    if (this.onError === null) {
+      throw new Error('JobsPromise: error handler must be set before adding promises')
+    }
     this.promises.add(promise)
+    const errorHandler = this.onError
     promise
       .catch((err: unknown) => {
-        this.onError?.(err)
+        errorHandler(err)
       })
       .finally(() => {
         this.promises.delete(promise)
